+++ packages/eyes-sdk-core/lib/utils/createFramesPaths.js
@@ -1,6 +1,6 @@
-function createFramesPaths(snapshot, path = []) {
-  const paths = snapshot.crossFramesXPaths
-    ? snapshot.crossFramesXPaths.map(selector => ({
+function createFramesPaths({snapshot, path = [], logger}) {
+  const paths = snapshot.crossFramesSelectors
+    ? snapshot.crossFramesSelectors.map(selector => ({
         path: path.concat(selector),
         parentSnapshot: snapshot,
       }))
@@ -8,12 +8,17 @@ function createFramesPaths(snapshot, path = []) {
 
   for (const frame of snapshot.frames) {
     if (frame.selector) {
-      paths.push(...createFramesPaths(frame, path.concat(frame.selector)))
+      paths.push(...createFramesPaths({snapshot: frame, path: path.concat(frame.selector), logger}))
     }
   }
 
+  logger.verbose(
+    `frames paths for ${snapshot.crossFramesSelectors}`,
+    paths.map(({path}) => path.join('-->')).join(' , '),
+  )
+
   delete snapshot.selector
-  delete snapshot.crossFramesXPaths
+  delete snapshot.crossFramesSelectors
 
   return paths
 }