+++ packages/eyes-sdk-core/test/unit/utils/createFramesPaths.spec.js
@@ -1,51 +1,64 @@
 const createFramesPaths = require('../../../lib/utils/createFramesPaths')
 const assert = require('assert')
+const {Logger} = require('../../../')
+
+const logger = new Logger(process.env.APPLITOOLS_SHOW_LOGS)
 
 describe('createFramesPaths', () => {
   it('should return an empty array when no cross frames exist', () => {
     const snapshot = {
       frames: [],
-      crossFramesXPaths: [],
+      crossFramesSelectors: [],
     }
-    const result = createFramesPaths(snapshot)
+    const result = createFramesPaths({snapshot, logger})
     assert.deepStrictEqual(result, [])
   })
 
   it('should create frame paths for cross origin frames', () => {
     const snapshot = {
       frames: [],
-      crossFramesXPaths: ['HTML[1]/BODY[1]'],
+      crossFramesSelectors: ['selector1'],
     }
-    const result = createFramesPaths(snapshot)
+    const result = createFramesPaths({snapshot, logger})
     assert.deepStrictEqual(result, [
       {
         parentSnapshot: {
           frames: [],
         },
-        path: ['HTML[1]/BODY[1]'],
+        path: ['selector1'],
       },
     ])
   })
 
   it('should create frame paths for frames that have cross origin frames', () => {
     const snapshot = {
+      cdt: 'top',
       frames: [
         {
+          cdt: 'parent',
           frames: [],
-          crossFramesXPaths: ['BODY[1]/IFRAME[1]'],
-          selector: 'BODY[1]',
+          crossFramesSelectors: ['selector1', 'selector2'],
+          selector: 'selector0',
         },
       ],
-      crossFramesXPaths: [],
+      crossFramesSelectors: [],
     }
 
-    const result = createFramesPaths(snapshot)
+    const result = createFramesPaths({snapshot, logger})
     assert.deepStrictEqual(result, [
       {
         parentSnapshot: {
+          cdt: 'parent',
           frames: [],
         },
-        path: ['BODY[1]', 'BODY[1]/IFRAME[1]'],
+        path: ['selector0', 'selector1'],
+      },
+      {
+        parentSnapshot: {
+          cdt: 'parent',
+          frames: [],
+        },
+        path: ['selector0', 'selector2'],
       },
     ])
   })