+++ packages/eyes-selenium/src/sdk.js
@@ -1,11 +1,11 @@
 const {EyesSDK} = require('@applitools/eyes-sdk-core')
 const VisualGridClient = require('@applitools/visual-grid-client')
-const spec = require('./SpecDriver')
+const spec = require('./spec-driver')
 const {version} = require('../package.json')
 
 module.exports = EyesSDK({
-  name: 'eyes.playwright',
-  version,
+  name: 'eyes.selenium',
+  version: `${version}--${process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION}`,
   spec,
   VisualGridClient,
 })