+++ packages/eyes-webdriverio-4/src/sdk.js
@@ -1,10 +1,10 @@
 const {EyesSDK} = require('@applitools/eyes-sdk-core')
 const VisualGridClient = require('@applitools/visual-grid-client')
-const spec = require('./SpecDriver')
+const spec = require('./spec-driver')
 const {version} = require('../package.json')
 
 module.exports = EyesSDK({
-  name: 'eyes-protractor',
+  name: 'eyes.webdriverio',
   version,
   spec,
   VisualGridClient,