+++ packages/eyes-webdriverio-5/src/sdk.js
@@ -1,6 +1,6 @@
 const {EyesSDK} = require('@applitools/eyes-sdk-core')
 const VisualGridClient = require('@applitools/visual-grid-client')
-const spec = require('./SpecDriver')
+const spec = require('./spec-driver')
 const {version} = require('../package.json')
 
 module.exports = EyesSDK({