+++ packages/sdk-shared/coverage-tests/custom/TestCheckDomSnapshotCrossOriginFrames_VG.spec.js
@@ -1,7 +1,6 @@
 const cwd = process.cwd()
 const path = require('path')
-const assert = require('assert')
-const spec = require(path.resolve(cwd, 'src/SpecDriver'))
+const spec = require(path.resolve(cwd, 'src/spec-driver'))
 const {testServer} = require('../../index')
 const {getEyes} = require('../../src/test-setup')
 const adjustUrlToDocker = require('../util/adjust-url-to-docker')