+++ packages/sdk-shared/coverage-tests/custom/TestIsDisabled.spec.js
@@ -2,7 +2,7 @@
 const cwd = process.cwd()
 const path = require('path')
 const {getEyes} = require('../../src/test-setup')
-const spec = require(path.resolve(cwd, 'src/SpecDriver'))
+const spec = require(path.resolve(cwd, 'src/spec-driver'))
 const {Target} = require(cwd)
 
 describe('Coverage Tests', () => {