+++ packages/eyes-webdriverio-5-service/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
+
+## 1.6.10 - 2020/10/1
+
 - remove yarn workspaces
 - updated to @applitools/eyes-webdriverio@5.22.2 (from 5.21.0)
+- updated to @applitools/eyes-webdriverio@5.23.1 (from 5.22.2)
 
 ## 1.6.9 - 2020/8/11